### 2018-04-02 #208 @richelbilderbeek
Submission: beautier
The purpose of `beautier` is to create a valid BEAST2 XML input file from its function arguments. In this way, a scientific pipeline using `BEAST2` can be fully scripted, instead of using `BEAUti`'s GUI.

```
Package: beautier
Title: 'BEAUti' from R
Version: 1.13
Author: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
Maintainer: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
Description: 'BEAST2' (<http://www.beast2.org>) is a widely used
  Bayesian phylogenetic tool, that uses DNA/RNA/protein data
  and many model priors to create a posterior of jointly estimated
  phylogenies and parameters.
  'BEAUti 2' (which is part of 'BEAST2') is a GUI tool
  that allows users to specify the many possible setups
  and generates the XML file 'BEAST2' needs to run.
  This package provides a way to create 'BEAST2' input
  files without active user input, but using
  R function calls instead.
License: GPL-3 | file LICENSE
    ape,
    geiger,
    seqinr,
    testit
    devtools,
    lintr,
```

https://github.com/richelbilderbeek/beautier

[e.g., "data extraction, because the package parses a scientific data file format"]

reproducibility, as it provides a way to script a configuration file, as an alternative to an error-prone (and tedious) GUI tool

Scientists that use the tool 'BEAST2' to do a Bayesian phylogenetic inference on DNA data.

No

Klaus Schliep, KlausVigo
Luke Harmon, lukejharmon
Jonathan Eastman, eastman
Joseph W. Brown, josephwb
