### 2018-04-02 #208 @richelbilderbeek
Note that:

 * The package is one of the three sub-packages of the packege `[babette](https://github.com/richelbilderbeek/babette)`. I intend to have all these packages on rOpenSci
 * There is already an article about `babette` in pre-print at [bioRxiv](https://www.biorxiv.org/content/early/2018/03/16/271866). The article (in the section 'Applications') is already submitted to Methods in Ecology and Evolution

I assume the better order would be to get the packages accepted by rOpenSci, then publish, but my supervisor does not see the added value of rOpenSci and thus wanted to get the article out ASAP.

As I do appreciate the feedback on my work and the ideology behind rOpenSci, I submit to rOpenSci nevertheless (in my spare time).
