### 2018-04-02 #208 @richelbilderbeek
This package is in [the process of being accepted by CRAN](https://github.com/richelbilderbeek/beautier/issues/33).
