### 2018-04-02 richelbilderbeek/beautier#34 @richelbilderbeek
Get on rOpenSci
Made [an issue at rOpenSci](https://github.com/ropensci/onboarding/issues/208)
