### 2018-04-02 #208 @noamross
Thanks for your submission, @richelbilderbeek! An initial question as we discuss this: can you tell us a bit about your logic for splitting up these functions into different packages? I can see reasons either way - doing One Thing Well and separating, or keeping together to maintain a common workflow.  Just curious how you see it as we think about the best way to use reviewer's time without overwhelming them, and also make sure that there is good feedback about the like use cases of using these packages in tandem.
