### 2018-04-02 #208 @richelbilderbeek
Thanks for you question, @noamross!

Sure, I'd be happy to help!

I hope this table will clear things up:

Step|Package|Tool
---|---|---
Configuration|`beautier`|BEAUti 2 (GUI)
Running|`beastier`|BEAST2 (command-line interface)
Parsing and analysing results|`tracerer`|Tracer (GUI)
---|---|---
 All of these|`babette`|None

As can be seen, each sub-package replaces a GUI or command-line tool (and is named after that tool). It follows the Do One Thing Well idea, and forced me in having no cycling dependencies. The bigger package, `babette` (a mix of the three subpackages' names), stitches these three packages together to do the full workflow.

What would be best not to overwhelm reviewers, I do not know. I think 3 out of 4 packages are easily reviewed, as here are the Single Line of Code counts (only files in `/R` folder, includes Doxygen lines):

package|SLOC
---|---
`beautier`|11695
`beastier`|896
`tracerer`|872
---|---
`babette`|272

If reviewing `babette` only/first would be best depends on -I think- on the administrative machinery behind the review process.

Whatever the choice or whatever I need to do, I'd be happy to have some feedback about my work.

Does this answer your question sufficiently?
