### 2018-04-04 #208 @noamross
Thanks, @richelbilderbeek! That helps.  After conferring, we've decided it would be best to submit all four as a single submission.  The joint workflow will be important and in the end its not that much more (relative) lines of code than `beautier`.

Since this will be a big review, we might assign a little more time than the standard 3 weeks to reviewers.  I also hope you can make things smoother by pre-running [`goodpractice::gp()`](https://github.com/MangoTheCat/goodpractice) on your packages to catch things that would likely be flagged in editor checks.

We also generally recommend going through review _before_ submitting to CRAN, as review often includes recommendations for breaking changes to the API, and we don't consider being on CRAN reason not to change these. Review usually also catches snafus that will hold up the CRAN submission process, too.

If that all makes sense, close this issue and open a new one for `babette` when you are ready, specifying in the text that the review is actually for all four packages.
