### 2018-04-04 #208 @richelbilderbeek
Happy my table helped @noamross !

 * About `goodpractive::gp()`: I already run that on my code, using Travis CI. Code coverage is 100%.
 * About CRAN: this has been put on hold, so I can now submit there after the rOpenSci review

I will follow your guidance, close this Issue and make a new one for `babette`.
